requirements.txt
Flask
requests
python-dotenv
pillow
waitress
reportlab
google-cloud-aiplatform